cloud_function_upload_gcs_to_bq/requirements.txt
functions-framework==3.*
google-cloud-storage==2.*
google-cloud-bigquery==3.*
ijson==3.1.4